fix - extract router & refactor
