+++ lib/app.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:home_task/screens/tiles.dart';
 
-import 'screens/color_list.dart';
-import 'screens/home.dart';
+import 'common/const.dart';
+import 'common/router.dart' as router;
 
 class App extends StatelessWidget {
   const App({Key? key}) : super(key: key);
@@ -11,12 +10,8 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Home Task',
-      initialRoute: '/',
-      routes: {
-        '/': (context) => const Home(),
-        '/tiles': (context) => const Tiles(),
-        '/list': (context) => const ColorList(),
-      },
+      initialRoute: homeScreenRoute,
+      onGenerateRoute: router.generateRoute,
       theme: ThemeData(
         appBarTheme: const AppBarTheme(
           centerTitle: false,
