+++ lib/common/const.dart
@@ -4,3 +4,8 @@ const title1 = '4 квадрата';
 const title2 = 'Цветной список';
 const title3 = 'Codelab';
 const title4 = 'Асинхронный запрос';
+
+//Routes
+const homeScreenRoute = '/';
+const tilesScreenRoute = 'tiles';
+const colorListScreenRoute = 'list';
